MVP-54 [파일정리] @bongbong811 작성했던 cg-set-address-screen 과 cg-search-address-screen 모두 _LEGACY 경로로 옮김. 이후, 파일명 변경 cg-set-address-temp-screen ➡️ cg-set-address-screen
