+++ app/screens/_LEGACY/legacy_test-map-screen/MyLocationMapMarker.tsx
